requirements.txt
flask
python-dotenv
pandas
requests
openai
beautifulsoup4
werkzeug
gunicorn
flask-httpauth
pytest
